separate out Trinidad::Daemon into trinidad/daemon.rb
