+++ spec/spec_helper.rb
@@ -6,14 +6,13 @@ rescue LoadError
   require 'rspec'
 end
 
-$:.unshift(File.dirname(__FILE__) + '/../lib')
-
-require 'trinidad_init_services/configuration'
-
-require 'java'
 require 'mocha'
-require 'fileutils'
 
 RSpec.configure do |config|
   config.mock_with :mocha
 end
+
+$:.unshift(File.dirname(__FILE__) + '/../lib')
+
+require 'trinidad_init_services'
+require 'trinidad/daemon'
